Add base api interactions class to backend service module and refactor services
