+++ app/services/scavengr_backend/user_scavenger_hunts.rb
@@ -1,11 +1,8 @@
 module ScavengrBackend
-  class UserScavengerHunts
-    attr_reader :base_url, :user
-
-    def initialize(current_user, username)
-      @base_url = 'https://scavengr-django.herokuapp.com'
-      @conn = Faraday.new(url: @base_url)
-      @user = current_user
+  class UserScavengerHunts < ScavengrBackend::BaseApiInteractions
+    def initialize(user, username)
+      super
+      @user = user
       @username = username
     end
 
@@ -14,15 +11,6 @@ module ScavengrBackend
       hunts = JSON.parse(response.body, symbolize_names: true)
       hunts.map do |hunt|
         ScavengerHunt.new(hunt)
-      end
-    end
-
-    private
-
-    def get(path)
-      @conn.get do |req|
-        req.url path
-        req.headers['Authorization'] = "Token #{user.token}"
       end
     end
   end
